Task #49: Address relay round-3 review
R3-1 (deep immutability & append binding): ValidatedLedger now stores a
private frozen LedgerModel in addition to the public Model view; LedgerAppend
and LedgerBuilder read the private snapshot only. A caller mutation of the
public Model via ImmutableCollectionsMarshal.AsArray no longer changes the
records used for continuation prefix equality, ordinal continuity, or
predecessor hash binding. A new test builds a bootstrap, mutates
bootstrap.Model.Records[0] via the marshal helper, and then successfully
appends a continuation — proving append uses the internal snapshot.

R3-2 (composite cause precedence): CountStructural now mirrors the canonical
writer exactly (17 header base props, 6 outcome base props, 5 changed-file
base props, finding always emits 8 base props). AssembleAndValidate checks
the 256 KiB canonical byte cap explicitly before invoking ParseAndValidate,
so a candidate above the 512 KiB raw cap surfaces as
canonical_byte_limit_exceeded rather than raw_byte_limit_exceeded. The
parser-failure wrap set is narrowed to the four frozen causes only —
interaction, property_count, array_length, canonical_byte_limit —
matching the refined issue section 9 contract.

R3-3 (null-safe preflight): HasInvalidUnicode accepts and rejects null so
any public API receiving a null caller-supplied string surfaces as
ledger_invalid_unicode instead of NullReferenceException.

R3-4 (long transition API): LedgerHeader.StateGeneration / LedgerEpoch /
PredecessorStateGeneration and every ExpectedTransition record now use
long per the refined API. LedgerCanonicalizer's header writer uses
WriteLong for these fields. Deserialization still uses TryGetInt32 because
the schema caps values well below 2^31, but the C# type contract now matches
the refined issue.

R3-major-5 (single source of truth for cross-check precedence):
AppendContinuation and CreateReset no longer perform their own predecessor
sanity checks; LedgerAppend.Validate{Continuation,Reset} is the single
source of truth for the cross-check order. The now-unused
ContinuationSanityCheck helper is removed.

R3-major-6 (real dual-fault mapper precedence tests):
LedgerSchemaMapperTests.ConstEnumViolation_PrecedesArrayLimitViolation and
UnknownField_PrecedesOtherClassifications now construct payloads that
simultaneously violate two rules and assert the higher-priority code fires.

R3-major-7 (PR body): rewritten to reflect the new public API, the empty
allowlist, the 86-test count, and the round-3 immutability contract.

+++ runtime/src/AgenticPrReview.Runtime/Ledger/LedgerCanonicalizer.cs
@@ -81,7 +81,7 @@ public static class LedgerCanonicalizer
             ["cacheConfigId"] = () => WriteString(buffer, h.CacheConfigId),
             ["headRepository"] = () => WriteString(buffer, h.HeadRepository),
             ["kind"] = () => WriteString(buffer, h.Kind),
-            ["ledgerEpoch"] = () => WriteInt(buffer, h.LedgerEpoch),
+            ["ledgerEpoch"] = () => WriteLong(buffer, h.LedgerEpoch),
             ["modelId"] = () => WriteString(buffer, h.ModelId),
             ["policyId"] = () => WriteString(buffer, h.PolicyId),
             ["predecessorLedgerSha256"] = () => WriteString(buffer, h.PredecessorLedgerSha256),
@@ -89,14 +89,14 @@ public static class LedgerCanonicalizer
             ["pullRequest"] = () => WriteInt(buffer, h.PullRequest),
             ["repository"] = () => WriteString(buffer, h.Repository),
             ["sessionEpoch"] = () => WriteString(buffer, h.SessionEpoch),
-            ["stateGeneration"] = () => WriteInt(buffer, h.StateGeneration),
+            ["stateGeneration"] = () => WriteLong(buffer, h.StateGeneration),
             ["templateId"] = () => WriteString(buffer, h.TemplateId),
             ["toolDefinitionId"] = () => WriteString(buffer, h.ToolDefinitionId),
             ["trustedExecutionDomain"] = () => WriteString(buffer, h.TrustedExecutionDomain),
             ["workflowIdentity"] = () => WriteString(buffer, h.WorkflowIdentity),
         };
-        if (h.PredecessorStateGeneration is int psg)
-            props["predecessorStateGeneration"] = () => WriteInt(buffer, psg);
+        if (h.PredecessorStateGeneration is long psg)
+            props["predecessorStateGeneration"] = () => WriteLong(buffer, psg);
         if (h.PredecessorManifestSha256 is string pms)
             props["predecessorManifestSha256"] = () => WriteString(buffer, pms);
         if (h.ResetReason is string rr)
